SMOODEV-tsdown: migrate fetch from tsup to tsdown
Drop-in build tooling swap — tsdown is the oxc-based replacement for tsup,
faster + actively maintained. The browser build's esbuild-plugin-alias
shim (which substitutes @smooai/logger Node entries with browser
variants) is replaced with @rollup/plugin-alias (rolldown-compatible)
+ deps.alwaysBundle for the same packages. tsdown emits
.cjs/.mjs/.d.cts/.d.mts (vs tsup's .js/.mjs/.d.ts); the exports map is
updated to match. fixedExtension is set explicitly on the browser config
so its output matches the core build's extension scheme.
